Bubble buildup. Bubble scaling with buildup. Bubble buildup with movement / velocity. Also made a temp camera functionality for rotating it without control rotation being used.
Bubble buildup component functionality for setting buildup amount and getting buildup amount.
Functions in the player interface for getting and updating the bubble buildup in case need to access it from external actors.
Bubble buildup affecting the scale of the actor but not the scale of the lil guy in the bubble since that is now using world scale instead of relative scale.
Bubble buildup changing using the event tick of the component with 0.01 seconds tick currently.
Multiple variables for the bubble buildup that can be changed to alter its behaviour for increasing and decreasing buildup intervals and speeds for the interp.
Bubble buildup updating with the velocity length of the owning actor of the component.