Alternative target buildup amount functionality which replaces the velocity checking amount. This is currently only used in the dashing but might be used elsewhere. Alternative max velocities for buildup based on being on the ground or not. Dashing stopping velocity with new buildup requirement. Different dash forces for vertical dashing and horizontal dashing.
Instead of the bubble buildup being solely the velocity based method it now can also used forced alternative buildup targets which are set through calling the update alternative target buildup function in the component.
This allows the lerping buildup to change for a certain amount of time to the most recent target amount.
This is used in the dashing to force the buildup to lerp currently to 0 when dashing is done.

Also alternative max velocities in the buildup component for whether the player is on the ground or not.
Also dashing checks for a certain amount of bubble buildup before negating current velocity so if lower doesn't stop velocity so bigger vertical dashes will stop the player's horizontal movement as well.

Need to do dash cooldown for horizontal dashing.
Also might do the indicator for the bubble buildup before sleeping.