Bubble Buildup indicator. Bubble buildup indicator material and functionality for updating its appearance based on bubble buildup amount. Separate stopping requirements for horizontal and vertical dashing now instead of just one.
Created a widget which is now on the player in world space and currently rotates to face the camera using a temp event with timer set on it.
This widget holds the image for the buildup indicator as well as a black border version slightly bigger than it.
Function to update the percentage in the widget using a dynamic material instance of the created buildup indicator material and the passed owning actor variable on begin play in the player.
All functionally works and looks pretty smooth obviously not aesthetically final yet but something to just show the buildup amount.

Changed the required stopping buildup amount in the dashing component to be separated into vertical and horizontal dash specific variables so can stop easier with the horizontal ones when dashing giving a bit more control back with the horizontal ones.

Used this tutorial to help create the buildup indicator material:https://youtu.be/60W4vcnfsg8?si=VqOkIlG4nG35lExQ