Camera. Dash tweaks for velocity.
New camera actor.
All camera functionality is currently for automatically following the player and using their mouse for rotation.

Camera actor is spawned in the player and is made a child of the player but using the absolute location and rotation and then set the view target with blend for the camera and the player controller.

Currently in the event tick for the camera with tick speed at 0.001 seconds. We get the axis values from the player interface which are the camera Y and camera Z variable values in the player which are set in the look input action functionality.
We use these values to set the relative rotation of the root for the camera with rotation speed variable and finterp.

Then setting the target arm length based on the velocity of the camera owner with various variables affecting the behaviour of this. The target arm length increases with the velocity with map range clamped. The Z velocity is divided by a division factor for the vector length leading into the map range clamped for the arm length.

Then we set the actor location of the camera actor. This is done based on the distance between the camera and the player for the XY values interping speed. Then for the speed of interping for the Z location of the camera we use a map range clamped and various variables for the current Z velocity of the player.