Dashing, jumping, landing camera shake (little).
Dash camera shake tied to the bubble amount when dashing so bigger dash = more camera shake.
Also the bubble buildup alternative target time is now tied to the bubble amount when dashing so bigger dash means a longer amount of time of the bubble amount being 0.

Functionality in on ground to do a small camera shake based on the velocity when being on the ground from not.

New functionality (possibly temporary) for finding and storing the previous 0.05 second's player velocity stored in a temp last velocity variable and a last velocity vector variable.

Jumping camera shake (a very small amount).

The camera shake is a singular instance as well.