Spectating and dead level cam functionality. Switching spectating targets. Dead ragdoll cam delay before level dead cam. Reset health now resets camera from spectating / dead
Player controller new graph for spectating and dead input actions and functionality.
New input actions and input mapping contexts for dead and spectating mainly switching between dead and spectating and then also switching spectating targets.
New actor for the dead level camera which is zoomed out and can see the entire level or the majority of it (this will slightly change in the next push).

Small amount of time between dying and going to the dead cam so can see the ragdolling bert.

The reset health function in the player health component sets back a lot of variables in the player controller when switching back to playing rather than dead / spectating.