CAMERA FIXES LETS GO OMG FINALLY LIVE CODING AND CODE COMPLETION IS ACTUALLY WORKING.
Created a custom blueprint function library for any custom functions we need for now.
Currently "made" a function for rinterping to a rotation value constantly but also using the shortest path functionality from rlerping using the code at the end of this discussion: https://forums.unrealengine.com/t/shortest-path-for-rinterp-to-node/374503/2

The custom function uses quaternion interpolation if shortest path is true which seems to have removed the glitchy camera movement.

This function is now used in the camera and dead camera actors for changing the relative rotation of their respective camera booms.
Had to slow down the mutliplier for camera movement speed in the player as well as lowering the maximum angle for Y camera movement.
Also altered the rotation speed values in these camera actors as well.