Dead camera view transforms instead of just rotations so can have different views in different areas that are toggleable between. Dead camera standard location for player input. Dead cam and spectating cam zooming with scroll wheel. Dead cam and spectating cam polishing.
The dead camera view that makes use of an array of transforms instead of just an array of rotations. This changes the functionality for finding the closest view from switching to dead view and from dying.
This also updates the event tick functionality in the dead view camera to lerp the current actor location to the target dead view location.
There is also therefore a standard vector for where the camera should be located when the player is able to rotate it whilst dead. This might change to use the location of the current dead view target location maybe.

Zoom functionality in the player controller for when spectating and when dead but with player input.
Using a zoom percentage delta.
Input action currently only on scroll wheel will need to do specific functionality /  a way to do it for controller since no scroll wheel on a controller duh.
0.5 percentage is the standard zoom.
Zoom functionality is based around the target arm length of the camera that is currently being used (dead cam or spectating cam which is the living camera).

Some other polishing changes including setting the camera location directly and immediately when going into spectating so its all still smooth even if the previous camera position was really far away when stopping spectating or dying.

Need to comment a lot of stuff for easier readability later and also need to make some more imaginary numbers into more obvious variables as well but will do that later.