Player Hitbox. Kill collision hazard setup. Player to player knockback manager and functionality. Altered the camera Z interp speed calculation to change if we are spectating so its way faster.
Player now has a sphere collision for their hitbox which is used for checking if they should die from kill volumes. This sphere is slightly smaller than their actual capsule collision and the visible bubble to give more leniency for the player.

New actor called bp_kill_volume which does as it says and uses a collision which only overlaps with the new player hitbox collision object type and then calls player death on the collided player health component.
New check in the death function to just ensure that the player is in fact alive when the death function is called.

New actor for managing player knockback when colliding with each other.
New collision on the player for the player to player collision which is slightly larger than the player's visible bubble and the player's collision.
Now when this overlaps with other player to player collisions the Player Knockback Activate function is called in the world's knockback manager passing the player who hit, player who was hit, velocities of both players and the sweep result's hit normal.

In the knockback manager actor we have an array of a new struct called player to player knockback info which just stores the 2 references for the players involved in the knockback, another array of floats with the player knockback times which correlates to the array of structs.
There are also some variables for the knockback multipliers depending on the player who hit and the player who got hit as well as the minimum time since the last duplicated knockback function call and finally the minimum combined velocity of both players for the knockback to actually trigger.

The functionality in the player knockback activate function makes use of checking the previous in world time that these two passed players collided with each other and checking if it is higher than a certain amount and if so we update the arrays by removing the found index and replacing it with the new game time in seconds and passed actors. Then we apply impulses to each player depending on who they are within the knockback function calls and their velocities altering the impulse amount.

Think there are still some improvements that could be made to this knockback but the majority of the underlying functionality is there. Sometimes the angle of the knockback is not really right I don't think.

Also added in some functionality in the main camera actor for checking if the target actor is the owning player when using the Z location interp speed and if it is not then we instead use a much higher interp speed so that switching between different spectating player's who are on different Z levels is much faster whilst remaining somewhat smooth.