Player Shadow Simple functionality using static mesh.
Simple actor created to update position and scale of a static mesh with very small Z scale which acts as the player shadow for now.
Want to do this a better way eventually so overhang over objects is better and better optimised etc but this works for now.
Spawned in the player and reference stored in the player.
New interface function for getting this player's shadow in the player interface.
Shadow is disabled on death and is not visible at all when on the ground.
Tick is disabled and renabled as needed so not unnecessarily running tick.