Camera modes. Camera spline based mode. Camera stationary mode. Stationary Mode.
New enum for camera modes.
These different camera modes will eventually be based on what the level is / mode for the level is. But not yet setup properly.

3 camera modes: 3rd person controllable, stationary, and spline based.

All functionality for these different modes in the camera actor.
Spline based makes use of 2 new actors: location spline and distance spline.
Stationary mode makes use of a new actor which just has a sphere, spring arm and camera in it which the camera uses for its transform and the target arm length of its spring arm / camera boom.
When player is dead just using the 3rd person controllable mode.

New struct for map info created but not setup yet.