Re-organisation. Main menu player controller, main menu player pawn, main menu base widget and actor containing world space widget. Setting up customisation UI display functionality.
New player controller specifically for being in the main menu.
New pawn actor for the player to control whilst being in the main menu so can keep unnecessary functionality out when not needed.
This pawn has simple camera movement and a widget interaction component to interact with various UI elements in the main menu.
Created a "base" main menu widget which currently has the main buttons to press and also created an actor go along with this with multiple splines to go between the different areas in the main menu with functionality to do so coordinated with the widget button presses.

Created some functionality for displaying and organising the customisation functionality with some more widgets.