More customisation functionality. Discovered mesh based sockets that can be used for customisation.
Customisation equipping and unequipping with auto unequipping when equipping a customisation with another with the same type already being equipped.

Might be able to use mesh sockets for individual positioning with standard naming for customisation placement and rotation for individual player models so don't have to make unique skeletons for each model.