Customisation character model rotator functionality with slider as its own widget. Multi Mesh customisations functionality setup for the hands and feet.
New struct and data table for customisations that use multiple meshes (currently setup for just two meshes but could change the functionality to do way more would just need more work).

New widget for customisation model rotating with a slider to change it (this is for keyboard so will need to do more functionality for controller support).