Scoreboard visuals and functionality making use of game state and player states for information. Pause menu very basic visuals but functionality there for now.
New input actions for toggling scoreboard and for pausing / unpausing.

Functionality for these input actions in the player controller in round one.

Created various widgets for the scoreboard so that it is modular to the amount of players in the game and their win counts.
Scoreboard also shows player names and will update an image for the player model that the player has using player info in the player states.

Pause menu activation sets input mode to UI only but then in the pause menu widget we check when P or ESC is pressed when pause menu has focus so can still use the pause menu buttons to close it when it is open but not allow other game oriented inputs whilst pause menu is active.

Not really much functionality to the pause menu other than pausing and resuming.

Will need to add some more functionality with scoreboard force closing possibly in certain circumstances.