Updated player info struct. Jump force orientation specific to surface angle jumping off.
Player info struct updated with map for customisations.

New functionality in the player event tick for updating a new sphere in the player that takes on the rotation of its surface using a box trace for world static, dynamic and physics actors making the rotation from Y impact normal and then rinterping to the new world rotation. This sphere's rotation is used for the up vector in the jump force calculation for impulse direction so jumping off of a sloped surface doesn't just send you directly up but instead opposite to the slope angle / direction.